rename stage metrics
